requirements/base.txt
# python modules
pillow
#psycopg2
requests
vectorformats
geopy

Django>=1.5

# django third party modules
easy_thumbnails
django-allauth
django-braces
django-compressor
django-pagination
django-positions
django-taggit
django-validated-file
south
